feat: Make extracted tck and feel testcases configs live in their own separate dirs
Test fir layout is following:

```
- tests-feel
    - TestCases <- Manually managed set of FEEL tests
	...
- tests-tck
    - TestCases <- Tests for complete TCK test suite
	...
```

`tests-feel/TestCases` are managed manually - usually test configs are
copied from `tests-tck` as soon as they are passing. Goal is to catch
regression bugs with these tests.

`tests-tck/TestCases` dir is managed via `make extract-testcases`. These
tests can be run via `make tests-tck`
